included exercise 1 in fibname.py file.
#Exercise 1 :
My name is Patrick, so the first and last letter of my name (P + K = 16 + 11) give the number  27. The 27th Fibonacci number is 196,418.

# Exercise 2 :
My surname is Meehan
The first letter M is number 77
The last letter n is number 110
Fibonacci number 187 is 538522340430300790495419781092981030533
ord() returns an integer representing the code of the character.